+++ nym-vpn-app/src-tauri/src/vpnd/feature_flags.rs
@@ -6,7 +6,6 @@ use ts_rs::TS;
 const KEY_QUIC: &str = "quic";
 const KEY_DOMAIN_FRONTING: &str = "domain_fronting";
 const KEY_ZKNYMS: &str = "zkNyms";
-const KEY_MIXNET_TUNING: &str = "mixnet_tuning";
 
 #[derive(Clone, Serialize, TS)]
 #[ts(export, export_to = "tauri.ts")]
@@ -15,7 +14,6 @@ pub struct FeatureFlags {
     pub quic: bool,
     pub domain_fronting: bool,
     pub zknym_credential: bool,
-    pub mixnet_tuning: bool,
 }
 
 impl From<lib::FeatureFlags> for FeatureFlags {
@@ -26,7 +24,6 @@ impl From<lib::FeatureFlags> for FeatureFlags {
                 .unwrap_or(false),
             zknym_credential: get_group_flag(&fflags, KEY_ZKNYMS, "credentialMode")
                 .unwrap_or(false),
-            mixnet_tuning: get_group_flag(&fflags, KEY_MIXNET_TUNING, "enabled").unwrap_or(false),
         }
     }
 }